revert back to  radiation_base.hpp + source_terms_single_group.hpp + source_terms_multi_group.hpp.
